+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ADDED:
 
+- [#4](https://github.com/woohoolabs/yang/pulls/4): Allow relationship's resource identifier objects to have a "meta" value. 
+
 CHANGED:
 
 REMOVED: